Expand Bag Pockets
The player can now hold all legally possible items at once. The Items pocket has 139 slots, while the Key Items pocket has 36 slots (3 more than needed to hold all legal FRLG items at once & the Old Sea Map).

All those items in the Items pocket is kind of absurd, so will be adding two more bag pockets to separate some of them out.

+++ src/item.c
@@ -121,7 +121,7 @@ void CopyItemName(u16 itemId, u8 * dest)
     }
 }
 
-s8 BagPocketGetFirstEmptySlot(u8 pocketId)
+s16 BagPocketGetFirstEmptySlot(u8 pocketId)
 {
     u16 i;
 
@@ -246,7 +246,7 @@ bool8 AddBagItem(u16 itemId, u16 count)
 {
     u8 i;
     u8 pocket;
-    s8 idx;
+    s16 idx;
 
     if (ItemId_GetPocket(itemId) == 0)
         return FALSE;
